Added Image Gallery
I don’t like the fact that the appearing image must be within the same
tag as the hover. I want to position them with flex box, but on the
opposite side in a separate container. Is it possible to hover over one
element and show a different one without the use of js?